+++ app/controllers/answers_controller.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+class AnswersController < ApplicationController
+  def create
+    plan = Plan.find(plan_id)
+    question = Question.find(question_id)
+
+    answer = Answer.new
+    answer.response = params[:response]
+    answer.question = question
+    answer.save
+
+    redirect_to plan_path(plan)
+  end
+
+  private
+
+  def plan_id
+    params[:plan_id]
+  end
+
+  def question_id
+    params[:question_id]
+  end
+end