+++ app/models/answer.rb
@@ -1,0 +1,4 @@
+class Answer < ApplicationRecord
+  self.implicit_order_column = "created_at"
+  belongs_to :question
+end