+++ app/services/get_contentful_question.rb
@@ -1,0 +1,8 @@
+class GetContentfulQuestion
+  def call
+    # TODO: Spec fixture only used for the first slice, needs a real API call
+    JSON.parse(
+      File.read("#{Rails.root}/spec/fixtures/contentful/radio-question-example.json")
+    )
+  end
+end