+++ spec/factories/answer.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :answer do
+    response { "Green" }
+  end
+end