+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog 1.0.0].
 - address rails-template TODO
 - any user can see a start page for the planning journey
 - use GOV.UK Frontend framework
+- users can see a basic form to start the planning journey sourced by a
+Contentful fixture
 
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/